Refactor survey reminder loop into repository Flow
Moved the periodic survey reminder checking logic out of `BellDashboardFragment` and into a continuous Flow in `SurveysRepository`. The repository now manages the SharedPreferences polling and cleanup on a background IO thread, while the fragment simply observes the results lifecycle-safely using `repeatOnLifecycle`.
