+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/BellDashboardFragment.kt
@@ -22,18 +22,16 @@ import androidx.recyclerview.widget.RecyclerView
 import dagger.hilt.android.AndroidEntryPoint
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.base.BaseDashboardFragment
 import org.ole.planet.myplanet.databinding.FragmentHomeBellBinding
 import org.ole.planet.myplanet.model.RealmSubmission
 import org.ole.planet.myplanet.model.RealmUser
+import org.ole.planet.myplanet.repository.SurveysRepository
 import org.ole.planet.myplanet.services.sync.ServerUrlMapper
 import org.ole.planet.myplanet.ui.courses.CoursesFragment
 import org.ole.planet.myplanet.ui.courses.TakeCourseFragment
@@ -52,7 +50,6 @@ class BellDashboardFragment : BaseDashboardFragment() {
     private var networkStatusJob: Job? = null
     private val viewModel: BellDashboardViewModel by viewModels()
     var user: RealmUser? = null
-    private var surveyReminderJob: Job? = null
     private var surveyListDialog: AlertDialog? = null
 
     @Inject
@@ -78,6 +75,7 @@ class BellDashboardFragment : BaseDashboardFragment() {
         setupNetworkStatusMonitoring()
         (activity as DashboardActivity?)?.supportActionBar?.hide()
         observeCompletedCourses()
+        observeSurveyReminders()
         viewLifecycleOwner.lifecycleScope.launch {
             val wasUserNull = user == null
             user = profileDbHandler.getUserModel()
@@ -153,8 +151,6 @@ class BellDashboardFragment : BaseDashboardFragment() {
 
     private fun checkPendingSurveys() {
         viewLifecycleOwner.lifecycleScope.launch {
-            if (checkScheduledReminders()) return@launch
-
             val preferences = requireActivity().getSharedPreferences(PREF_SURVEY_REMINDERS, 0)
             val lastShown = preferences.getLong(KEY_LAST_SURVEY_DIALOG_SHOWN, 0L)
             if (System.currentTimeMillis() - lastShown < SURVEY_DIALOG_INTERVAL_MS) return@launch
@@ -170,8 +166,6 @@ class BellDashboardFragment : BaseDashboardFragment() {
                 )
                 val surveyTitles = submissionsRepository.getSurveyTitlesFromSubmissions(pendingSurveys)
                 showSurveyListDialog(pendingSurveys, title, surveyTitles)
-            } else {
-                checkScheduledReminders()
             }
         }
     }
@@ -237,64 +231,30 @@ class BellDashboardFragment : BaseDashboardFragment() {
             putLong("reminder_time_$surveyIds", reminderTime)
                 .putString("reminder_surveys_$surveyIds", surveyIds)
         }
-
-        startReminderCheck()
     }
 
-    private fun startReminderCheck() {
-        surveyReminderJob?.cancel()
-        surveyReminderJob = lifecycleScope.launch {
-            while (isActive) {
-                checkScheduledReminders()
-                delay(60000)
-            }
-        }
-    }
-
-    private suspend fun checkScheduledReminders(): Boolean {
-        val currentTime = System.currentTimeMillis()
-
-        val (remindersToShow, remindersToRemove, preferences) = withContext(Dispatchers.IO) {
-            val prefs = requireActivity().getSharedPreferences(PREF_SURVEY_REMINDERS, 0)
-            val toShow = mutableListOf<String>()
-            val toRemove = mutableListOf<String>()
-            for (entry in prefs.all) {
-                if (entry.key.startsWith("reminder_time_")) {
-                    val surveyIds = entry.key.removePrefix("reminder_time_")
-                    val reminderTime = prefs.getLong(entry.key, 0)
-                    if (reminderTime <= currentTime) {
-                        toShow.add(surveyIds)
-                        toRemove.add(surveyIds)
-                    }
+    private fun observeSurveyReminders() {
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                surveysRepository.dueRemindersFlow().collect { ids ->
+                    handleDueReminders(ids)
                 }
             }
-            Triple(toShow, toRemove, prefs)
         }
+    }
 
+    private suspend fun handleDueReminders(remindersToShow: List<String>) {
         for (surveyIds in remindersToShow) {
             val surveyIdList = surveyIds.split(",").filter { it.isNotBlank() }
-            if (surveyIdList.isEmpty()) {
-                continue
-            }
+            if (surveyIdList.isEmpty()) continue
             val submissions = submissionsRepository.getSubmissionsByIds(surveyIdList)
             val submissionsById = submissions.associateBy { it.id }
-            val pendingSurveys = surveyIdList.mapNotNull { submissionsById[it] }
-                .filter { it.status == "pending" }
+            val pendingSurveys = surveyIdList.mapNotNull { submissionsById[it] }.filter { it.status == "pending" }
 
             if (pendingSurveys.isNotEmpty()) {
                 showPendingSurveysReminder(pendingSurveys)
             }
         }
-
-        preferences.edit {
-            for (surveyIds in remindersToRemove) {
-                remove("reminder_time_$surveyIds")
-                remove("reminder_surveys_$surveyIds")
-            }
-        }
-
-        return remindersToShow.isNotEmpty()
-
     }
 
     private fun showPendingSurveysReminder(pendingSurveys: List<RealmSubmission>) {
@@ -463,7 +423,6 @@ class BellDashboardFragment : BaseDashboardFragment() {
         surveyListDialog?.dismiss()
         surveyListDialog = null
         networkStatusJob?.cancel()
-        surveyReminderJob?.cancel()
         super.onDestroyView()
         _binding = null
     }
